+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
+## v0.2.6
+- Fixed a bug in vdom rendering of children
+- Improved vdom logic
+
 ## v0.2.5
 - Attributes and Events now can use `At` and `Ev` enums
 - Routing overhauled; modelled after react-reason. Cleaner syntax, and more flexible.
 - Input, Textarea, and Select elements are now "controlled" - they always
 stay in sync with the model.
 - index.html file updated in examples and quickstart to use relative paths,
-which fixes some routing problems
+which fixes landing-page routing
 
 ## v0.2.4
 - Changed render func to use a new pattern (Breaking)